+++ test/which_test.js
@@ -9,53 +9,61 @@ const {Finder, which} = require('../lib');
 describe('which()', () => {
   let options = {path: 'test/fixtures'};
 
-  it('should return the path of the `executable.cmd` file on Windows', done => {
-    which('executable', false, options).subscribe(
-      executable => {
-        if (!Finder.isWindows) done(new Error('Error not thrown.'));
-        else expect(executable.endsWith('\\test\\fixtures\\executable.cmd')).to.be.true;
-      },
-      err => done(Finder.isWindows ? new Error(err.message) : null),
-      done
-    );
+  it('should return the path of the `executable.cmd` file on Windows', async () => {
+    try {
+      let executable = await which('executable', false, options);
+      if (Finder.isWindows) expect(executable.endsWith('\\test\\fixtures\\executable.cmd')).to.be.true;
+      else expect(true).to.not.be.ok;
+    }
+
+    catch (err) {
+      if (Finder.isWindows()) expect(true).to.not.be.ok;
+      else expect(true).to.be.ok;
+    }
   });
 
-  it('should return all the paths of the `executable.cmd` file on Windows', done => {
-    which('executable', true, options).subscribe(
-      executables => {
-        if (!Finder.isWindows) done(new Error('Error not thrown.'));
-        else {
-          expect(executables).to.be.an('array').and.have.lengthOf(1);
-          expect(executables[0]).to.be.a('string').and.match(/\\test\\fixtures\\executable\.cmd$/);
-        }
-      },
-      err => done(Finder.isWindows ? new Error(err.message) : null),
-      done
-    );
+  it('should return all the paths of the `executable.cmd` file on Windows', async () => {
+    try {
+      let executables = await which('executable', true, options);
+      if (!Finder.isWindows) expect(true).to.not.be.ok;
+      else {
+        expect(executables).to.be.an('array').and.have.lengthOf(1);
+        expect(executables[0]).to.be.a('string').and.match(/\\test\\fixtures\\executable\.cmd$/);
+      }
+    }
+
+    catch (err) {
+      if (Finder.isWindows()) expect(true).to.not.be.ok;
+      else expect(true).to.be.ok;
+    }
   });
 
-  it('should return the path of the `executable.sh` file on POSIX', done => {
-    which('executable.sh', false, options).subscribe(
-      executable => {
-        if (Finder.isWindows) done(new Error('Error not thrown.'));
-        else expect(executable.endsWith('/test/fixtures/executable.sh')).to.be.true;
-      },
-      err => done(Finder.isWindows ? null : new Error(err.message)),
-      done
-    );
+  it('should return the path of the `executable.sh` file on POSIX', async () => {
+    try {
+      let executable = await which('executable.sh', false, options);
+      if (Finder.isWindows) expect(true).to.not.be.ok;
+      else expect(executable.endsWith('/test/fixtures/executable.sh')).to.be.true;
+    }
+
+    catch (err) {
+      if (Finder.isWindows()) expect(true).to.be.ok;
+      else expect(true).to.not.be.ok;
+    }
   });
 
-  it('should return all the paths of the `executable.sh` file on POSIX', done => {
-    which('executable.sh', true, options).subscribe(
-      executables => {
-        if (Finder.isWindows) done(new Error('Error not thrown.'));
-        else {
-          expect(executables).to.be.an('array').and.have.lengthOf(1);
-          expect(executables[0]).to.be.a('string').and.match(/\/test\/fixtures\/executable\.sh$/);
-        }
-      },
-      err => done(Finder.isWindows ? null : new Error(err.message)),
-      done
-    );
+  it('should return all the paths of the `executable.sh` file on POSIX', async () => {
+    try {
+      let executables = await which('executable.sh', true, options);
+      if (Finder.isWindows) expect(true).to.not.be.ok;
+      else {
+        expect(executables).to.be.an('array').and.have.lengthOf(1);
+        expect(executables[0]).to.be.a('string').and.match(/\/test\/fixtures\/executable\.sh$/);
+      }
+    }
+
+    catch (err) {
+      if (Finder.isWindows()) expect(true).to.be.ok;
+      else expect(true).to.not.be.ok;
+    }
   });
 });